+++ include/omw/utility.h
@@ -1,7 +1,7 @@
 /*
 author          Oliver Blaser
-date            25.01.2022
-copyright       MIT - Copyright (c) 2022 Oliver Blaser
+date            05.03.2022
+copyright       MIT - Copyright (c) 2023 Oliver Blaser
 */
 
 #ifndef IG_OMW_UTILITY_H
@@ -59,6 +59,51 @@ namespace omw
     /*! \addtogroup grp_utility_gpUtil
     * @{
     */
+
+    class Base_Nullable
+    {
+    public:
+        Base_Nullable() : m_isNull(true) {}
+        explicit Base_Nullable(bool isNull) : m_isNull(isNull) {}
+        virtual ~Base_Nullable() {}
+
+        bool isNull() const { return m_isNull; }
+        
+        virtual void makeNull() { m_isNull = true; }
+
+    protected:
+        void setIsNull(bool isNull) { m_isNull = isNull; }
+
+    private:
+        bool m_isNull;
+    };
+
+    template <typename T>
+    class Nullable : public Base_Nullable
+    {
+    public:
+        using value_type = T;
+        using reference = value_type&;
+        using const_reference = const value_type&;
+
+    public:
+        Nullable() : Base_Nullable(), m_value() {}
+        Nullable(const_reference value) : Base_Nullable(false), m_value(value) {}
+        virtual ~Nullable() {}
+
+        const_reference get(const_reference fallback) const { return (isNull() ? fallback : m_value); }
+        void set(const_reference value) { setIsNull(false); m_value = value; }
+
+        virtual void free() { m_value = value_type(); }
+        virtual void makeNull() { free(); Base_Nullable::makeNull(); }
+
+        explicit operator value_type() const { return m_value; }
+
+    protected:
+        value_type m_value;
+    };
+
+
 
     inline void toggle(bool& value) { value = !value; }
     inline void toggle(int& value) { value = (value ? 0 : 1); }