+++ release-notes.md
@@ -11,6 +11,7 @@
 New
  - `omw::stoz()`
  - `omw::string::reverse()` and `reversed()`
+ - `omw::Base_Nullable` and `omw::Nullable<T>`
  - `omw::io::FileInterface_Base::openRead()`
  - `omw::io::FileInterface_Base::openWrite()`
  - `omw::io::SerialPort`